handle selected 0 items

+++ server.js
@@ -147,13 +147,17 @@ app.post('/update-purchase_time', (req, res) => {
 });
 
 app.post('/icons-to-home', (req, res) => {
-    console.log(req.body.si);
-    console.log(req.session.user_id);
-    var selectedItem = req.body.si;
-    con.query('DELETE FROM user_data WHERE user_id = ?', [req.session.user_id], (err, results, fields) => {
-        if (err) throw err;
-        console.log('deleted all rows for current user...')
-    })
+    console.log(`req.body.si ${req.body.si}`);
+    if (req.body.si) {
+        var selectedItem = req.body.si.map(x => {
+            return parseInt(x);
+        });
+    } else {
+        con.query('DELETE FROM user_data WHERE user_id = ?', [req.session.user_id], (err, results, fields) => {
+            if (err) throw err;
+            console.log('deleted all rows for current user...')
+        })
+    }
 
     for (let i in selectedItem) {
         con.query('INSERT INTO user_data (user_id, food_id, added_to_home) VALUES (?,?,1)', [req.session.user_id, selectedItem[i]], (err, results, fields) => {
@@ -161,7 +165,7 @@ app.post('/icons-to-home', (req, res) => {
             console.log(`Added food_id of ${selectedItem[i]} into user_data table, current user : ${req.session.user_id}...`);
         })
     }
-    res.redirect('/homedash');
+    // res.redirect('/homedash');
 });
 
 app.delete('/remove-food', (req, res) => {
